Many changes made to make package acceptable to CRAN.
Many changes made to make the package acceptable to CRAN. The code is the code that was submitted and accepted. There is a bug in summary.mediation.R and print.mediation.R because the summary function doesn't work on dx.wts objects. I have a fix that I will make once GitHub and my PC are in sync